requirements.txt
opencv
numpy
math
matplotlib.pyplot 
pickle
sys
time